Tweaks to make it behave
Without moving the 'opens' keyword to the position it is now found
in (in java.g) there are problems with reducing QualifiedName -
it looks to be due to parser generation changing the ids for some
key tokens (like identifier). If those don't move, you are ok, if
they move then you get problems in reduction when trying to use it.

When you have reduction problems you can't process even simple
things like a program that is simply 'import java.util.List;'