Fix APT handling for some transformers
Eclipse appears to be closing the classloader used to
load the classes related to a particular APT before
the final round of processing is done. This means if
the final round needs to load more classes it fails with
a NoClassDefFound. These changes defer closing the
classloader until all processing has been done.